+++ website/src/components/gallery/ShowcaseCard/index.tsx
@@ -66,7 +66,9 @@ function ShowcaseCard({ user }: { user: User }): JSX.Element {
   const tags = [
     ...user.tags,
     ...(user.languages || []),
+    ...(user.frameworks || []),
     ...(user.azureServices || []),
+    ...(user.IaC || []),
   ];
   const source = user.source;
   const star = useBaseUrl("/img/Sparkle.svg");